Refactorizacion terminada
