+++ components/forms/miniForms/FormLeadership.jsx
@@ -1,55 +1,54 @@
 // Layout
-import InputGroup from "./layout/InputGroup"
+import InputGroup from "../layout/InputGroup";
 // Components
-import { DateForm, InputForm, TextareaForm } from "@/components/component/InputForm"
-import { Button } from "../ui/button";
-//  Esquemas
-import { experienceSchema } from "@/utils/schemas/userInfo.schema"
-//  Hooks
-import { useForm } from "@/utils/hooks/useForm"
+import { DateForm, InputForm, TextareaForm } from "../../component/InputForm";
+import { Button } from "../../ui/button";
+// Esquemas
+import { leadershipSchema } from "@/utils/schemas/userInfo.schema";
+// Hooks
+import { useForm } from "@/utils/hooks/useForm";
 // Utilidades
-import { dateToday } from "@/utils/helper/time-handlers/today"
+import { dateToday } from "@/utils/helper/time-handlers/today";
 
-const FormExperience = ({ onClick }) => {
 
-    // Estado Inicial del formulario
+const FormLeadership = ({ onClick }) => {
+
+    // Estado inicial del formulario
     const initialState = {
         organization: '',
         location: '',
-        position: '',
+        role: '',
         startDate: '',
         endDate: '',
-        description: ''
+        achievements: ''
     }
 
     // Hook para el control del formulario
-    const { errors, formData, handleInput, add } = useForm(experienceSchema, initialState, onClick)
+    const { errors, formData, handleInput, add } = useForm(leadershipSchema, initialState, onClick)
 
     return (
-        <InputGroup title={"Experiencia"}>
+        <InputGroup title={"Liderazgo y Actividades"}>
             <div className="grid grid-cols-2 gap-4">
-
                 <InputForm
                     name='organization'
-                    label='Empresa o Organización'
-                    maxLenght={50}
+                    label='Organización'
                     value={formData.organization}
                     _errors={errors.organization?._errors}
                     onChange={handleInput}
-                    placeholder={"Nombre de la Organizacion/Empresa"}
+                    maxLenght={50}
                 />
                 <InputForm
-                    name='position'
-                    label='Posición'
-                    maxLenght={50}
-                    value={formData.position}
-                    _errors={errors.position?._errors}
+                    name='role'
+                    label='Rol'
+                    value={formData.role}
+                    _errors={errors.role?._errors}
                     onChange={handleInput}
-                    placeholder={"Título del empleo"}
+                    maxLenght={50}
+                    placeholder={"Ingresa tu Rol/Posición"}
+                    title={"Cual fue tu rol/posición, que hiciste?"}
                 />
             </div>
             <div className="grid grid-cols-3 gap-2">
-
                 <DateForm
                     name='startDate'
                     label='Fecha de Inicio'
@@ -61,7 +60,7 @@ const FormExperience = ({ onClick }) => {
                 />
                 <DateForm
                     name='endDate'
-                    label='Fecha de Fin'
+                    label='Fecha de Finalización'
                     value={formData.endDate}
                     _errors={errors.endDate?._errors}
                     onChange={handleInput}
@@ -71,10 +70,10 @@ const FormExperience = ({ onClick }) => {
                 <InputForm
                     name='location'
                     label='Lugar'
-                    maxLenght={50}
                     value={formData.location}
                     _errors={errors.location?._errors}
                     onChange={handleInput}
+                    maxLenght={50}
                 />
             </div>
             <div className="grid grid-cols-1">
@@ -84,13 +83,12 @@ const FormExperience = ({ onClick }) => {
                     value={formData.description}
                     _errors={errors.description?._errors}
                     onChange={handleInput}
-                    rows={4}
-                    placeholder={"Descripción detallada de tu utilidad en la compañia"}
+                    placeholder={"Describe extensivamente como fue tu participación en este evento o que logros obtuviste"}
                 />
             </div>
-            <Button onClick={add} className="border hover:bg-slate-50 hover:text-black" >Añadir Experiencia</Button>
+            <Button onClick={add} className="border hover:bg-slate-50 hover:text-black" >Añadir Liderazgo</Button>
         </InputGroup>
     )
 }
 
-export default FormExperience
+export default FormLeadership;
